--- packages/shopkeeper/src/middlewares/stripe_webhook_middleware.ts
@@ -1,27 +1,0 @@
-import { HttpContext } from '@adonisjs/core/http'
-import { NextFn } from '@adonisjs/core/types/http'
-import shopkeeper from '../../services/shopkeeper.js'
-
-export default class StripeWebhookMiddleware {
-  async handle({ request }: HttpContext, next: NextFn) {
-    const sig = request.header('stripe-signature')
-    const body = request.raw()
-
-    if (!body || !sig) {
-      throw new Error('') // TODO: Error
-    }
-
-    const valid = shopkeeper.stripe.webhooks.signature.verifyHeader(
-      body,
-      sig,
-      shopkeeper.config.webhook.secret!, // TODO: Error
-      shopkeeper.config.webhook.tolerance
-    )
-
-    if (!valid) {
-      throw new Error('') // TODO: Error
-    }
-
-    await next()
-  }
-}
